Move LLM types
