+++ src/containers/FormikForm/MetaDataField.tsx
@@ -45,8 +45,8 @@ import { FormContent } from "../../components/FormikForm";
 import PlainTextEditor from "../../components/SlateEditor/PlainTextEditor";
 import { textTransformPlugin } from "../../components/SlateEditor/plugins/textTransform";
 import { AI_ACCESS_SCOPE, DRAFT_ADMIN_SCOPE } from "../../constants";
+import { MetaDescriptionVariables, SummaryVariables } from "../../interfaces";
 import { useDraftSearchTags } from "../../modules/draft/draftQueries";
-import { MetaDescriptionVariables, SummaryVariables } from "../../server/llmApiTypes";
 import { inlineContentToEditorValue } from "../../util/articleContentConverter";
 import useDebounce from "../../util/useDebounce";
 import { useSession } from "../Session/SessionProvider";
