+++ src/interfaces.ts
@@ -205,3 +205,65 @@ export interface SearchParams {
   "filter-inactive"?: boolean;
   includeCopyrighted?: boolean;
 }
+
+export type PromptType = PromptVariables["type"];
+
+export type PromptVariables =
+  | SummaryVariables
+  | AltTextVariables
+  | AlternativePhrasingVariables
+  | MetaDescriptionVariables
+  | ReflectionVariables;
+
+export interface SummaryVariables {
+  type: "summary";
+  text: string;
+  title: string;
+}
+
+export interface AltTextVariables {
+  type: "altText";
+  image: {
+    fileType: string;
+    base64: string;
+  };
+}
+
+export interface AlternativePhrasingVariables {
+  type: "alternativePhrasing";
+  text: string;
+  excerpt: string;
+}
+
+export interface MetaDescriptionVariables {
+  type: "metaDescription";
+  text: string;
+  title: string;
+}
+
+export interface ReflectionVariables {
+  type: "reflection";
+  text: string;
+}
+
+const promptTypes: PromptType[] = [
+  "summary",
+  "altText",
+  "alternativePhrasing",
+  "metaDescription",
+  "reflection",
+] as const;
+
+export const isPromptType = (type: string): type is PromptType => promptTypes.includes(type as PromptType);
+
+export type PromptPayload<T extends PromptVariables> = T & {
+  language: string;
+  role?: string;
+  instructions?: string;
+  max_tokens?: number;
+};
+
+export interface DefaultPrompts {
+  role: string;
+  instructions: string;
+}
